24EG505E02,01,04_Meeting Summary Generator/requirements.txt
groq
python-docx
pydub
python-dotenv